add ColorSchemeSelector helper component
