+++ src/app/js/formats/cartography/CartographyAdmin.js
@@ -1,49 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import SelectField from 'material-ui/SelectField';
-import MenuItem from 'material-ui/MenuItem';
 import translate from 'redux-polyglot/translate';
-import {
-    schemeBlues,
-    schemeOrRd,
-    schemeBuGn,
-    schemeBuPu,
-    schemeGnBu,
-    schemeGreys,
-    schemeOranges,
-    schemeGreens,
-    schemePuBu,
-    schemePuBuGn,
-    schemePuRd,
-    schemeRdPu,
-    schemeYlGn,
-    schemeYlGnBu,
-    schemeYlOrBr,
-    schemeYlOrRd,
-} from 'd3-scale-chromatic';
-import { scaleQuantize } from 'd3-scale';
+import { schemeBlues, schemeOrRd } from 'd3-scale-chromatic';
 import TextField from 'material-ui/TextField';
 
-import ColorScalePreview from './ColorScalePreview.js';
-
-const schemes = [
-    schemeBlues[9],
-    schemeOrRd[9],
-    schemeBuGn[9],
-    schemeBuPu[9],
-    schemeGnBu[9],
-    schemeGreys[9],
-    schemeOranges[9],
-    schemeGreens[9],
-    schemePuBu[9],
-    schemePuBuGn[9],
-    schemePuRd[9],
-    schemeRdPu[9],
-    schemeYlGn[9],
-    schemeYlGnBu[9],
-    schemeYlOrBr[9],
-    schemeYlOrRd[9],
-];
+import ColorSchemeSelector from '../../lib/components/ColorSchemeSelector';
 
 import { polyglot as polyglotPropTypes } from '../../propTypes';
 
@@ -134,46 +95,18 @@ class CartographyAdmin extends Component {
                     underlineFocusStyle={{ borderColor: defaultColor }}
                     value={defaultColor}
                 />
-                <SelectField
-                    floatingLabelText={polyglot.t('color_scheme')}
+                <ColorSchemeSelector
+                    label={polyglot.t('color_scheme')}
                     onChange={this.setColorScheme}
                     style={styles.input}
-                    value={colorScheme.join(',')}
-                >
-                    {schemes.map(value => (
-                        <MenuItem
-                            key={value}
-                            value={value.join(',')}
-                            primaryText={
-                                <ColorScalePreview
-                                    colorScale={scaleQuantize()
-                                        .domain([0, 100])
-                                        .range(value)}
-                                />
-                            }
-                        />
-                    ))}
-                </SelectField>
-                <SelectField
-                    floatingLabelText={polyglot.t('hover_color_scheme')}
+                    value={colorScheme}
+                />
+                <ColorSchemeSelector
+                    label={polyglot.t('hover_color_scheme')}
                     onChange={this.setHoverColorScheme}
                     style={styles.input}
-                    value={hoverColorScheme.join(',')}
-                >
-                    {schemes.map(value => (
-                        <MenuItem
-                            key={value}
-                            value={value.join(',')}
-                            primaryText={
-                                <ColorScalePreview
-                                    colorScale={scaleQuantize()
-                                        .domain([0, 100])
-                                        .range(value)}
-                                />
-                            }
-                        />
-                    ))}
-                </SelectField>
+                    value={hoverColorScheme}
+                />
             </div>
         );
     }
