+++ src/app/js/lib/components/ColorScalePreview.js
